requirements.txt
aiohttp==3.8.3
aiolimiter==1.0.0
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
certifi==2020.6.20
chardet==3.0.4
charset-normalizer==2.1.1
click==8.0.0
dnacentersdk==2.0.2
fastjsonschema==2.16.2
Flask==2.2.2
frozenlist==1.3.3
future==0.18.2
idna==2.10
itsdangerous==2.1.0
Jinja2==3.1.0
MarkupSafe==2.1.1
meraki==1.27.0
multidict==6.0.4
PyJWT==1.7.1
python-dotenv==0.14.0
requests==2.24.0
requests-toolbelt==0.9.1
urllib3==1.25.10
webexteamssdk==1.6
Werkzeug==2.2.2
yarl==1.8.2